fix: golden 内审 14 例修复（伪缺陷 3 / 场景标错 6 / 存疑 5）
内审记录 reports/audit-internal-2026-09-03.md。
- 伪缺陷重写：c12 环改 64 项（cwe-125）、r07 清零循环改 size_t 域
  （4GiB 越界写）、r09 删错误路径置 NULL（真 double-free）
- 场景标错：c01b/c03/t01/r02/r05 改 cwe-787；r12 改 cwe-125，按 gcc
  实测纠正审计方案（16u 反造伪缺陷），源码未动、机理重写为缺下界
  检查
- 存疑修复：c15 改 _Atomic+acquire/release（dequeue 取模安全点）；
  c21 GBUF_N 改 128 使回绕漏网真实越界；r08 补 join happens-before
  契约；r11 活动树加 r11_get_ok 作 FP 探针；r13 EV_A/B 填真实
  handler 仅 EV_C 缺失
全部经 ASan/pthread 构造性验证。涉及基线数字待 18:00 后统一重跑。

+++ cases/defect/r07-alloc-size-wrap/src/alloc.c
@@ -12,7 +12,7 @@ void *r07_alloc(uint32_t n, uint32_t size)
     if (p == NULL) {
         return NULL;
     }
-    for (uint32_t i = 0; i < n * size; i++) {   /* 逐字节清零 */
+    for (size_t i = 0; i < (size_t)n * size; i++) {   /* 逐字节清零 */
         p[i] = 0;
     }
     return p;
